--- src/core/Source1.cpp
@@ -1,5 +1,0 @@
-#include "../include/Source1.h"
-
-int doubleMe(int x) {
-	return x * x;
-}
