--- src/core/coolnumber.cpp
@@ -1,9 +1,0 @@
-//
-// Created by Konstantin Grigoriev on 11/10/2019.
-//
-
-#include "../include/coolnumber.h"
-
-int getCoolNumber() {
-	return 42;
-}
