--- src/include/EnumToString.h
@@ -1,8 +1,0 @@
-#ifndef MINI_CPP_COMPILER_ENUMTOSTRING_H
-#define MINI_CPP_COMPILER_ENUMTOSTRING_H
-
-#include "../include/Token.h"
-
-string enToStr(LexemType type);
-
-#endif
