--- src/include/Source1.h
@@ -1,7 +1,0 @@
-
-#ifndef MINI_CPP_COMPILER_SOURCE1_H
-#define MINI_CPP_COMPILER_SOURCE1_H
-
-int doubleMe(int x);
-
-#endif
