--- src/include/coolnumber.h
@@ -1,10 +1,0 @@
-//
-// Created by Konstantin Grigoriev on 11/10/2019.
-//
-
-#ifndef MINI_CPP_COMPILER_COOLNUMBER_H
-#define MINI_CPP_COMPILER_COOLNUMBER_H
-
-int getCoolNumber();
-
-#endif //MINI_CPP_COMPILER_COOLNUMBER_H
