+++ src/main.cpp
@@ -1,12 +1,9 @@
 #include <iostream>
-#include "include/coolnumber.h"
 #include "include/Token.h"
-#include "include/EnumToString.h"
 
 
 int main() {
 	std::cout << "Hello, World!" << std::endl;
-	std::cout << getCoolNumber() << std::endl;
 	LexemType fru = LexemType::chr;
 	Token token('a');
 	cout << token.value() << endl;
